final project

+++ app/src/main/java/com/example/divasalon/models/User.kt
@@ -4,7 +4,7 @@ import com.google.gson.annotations.SerializedName
 data class User(
     @SerializedName("customer_name") val fullName: String,
     @SerializedName("email") val email: String,
-    @SerializedName("phone") val phone: Int,
+    @SerializedName("phone") val phone: Long,
     @SerializedName("password") val password: String
 )
